add build option to run script to avoid re-building each time

+++ src/run_clx.sh
@@ -3,15 +3,21 @@
 module load cmake/3.18.3
 module load intel/compiler/64/2020/19.1.3
 
+build=true
+
+if $build; then
 rm -rf build_*
+fi
 
 export OMP_NUM_THREADS=40
 export OMP_PLACES=cores
 export OMP_PROC_BIND=true
 
 # Build OpenMP
+if $build; then
 cmake -Bbuild_omp -H. -DMODEL=OpenMP -DCMAKE_CXX_COMPILER=icpc
 cmake --build build_omp --parallel
+fi
 
 if [ -f ./build_omp/Reduced ]; then
   for b in dot complex_sum complex_sum_soa complex_min field_summary describe; do
@@ -24,6 +30,7 @@ fi
 
 
 # Build Kokkos
+if $build; then
 if [ ! -d kokkos-3.4.01 ]; then
   wget https://github.com/kokkos/kokkos/archive/refs/tags/3.4.01.tar.gz
   tar xf 3.4.01.tar.gz
@@ -32,6 +39,7 @@ fi
 
 cmake -H. -Bbuild_kokkos -DMODEL=Kokkos -DKOKKOS_SRC=kokkos-3.4.01 -DKokkos_ENABLE_OPENMP=On -DKokkos_ARCH_SKX=On -DCMAKE_CXX_COMPILER=icpc
 cmake --build build_kokkos --parallel
+fi
 
 
 if [ -f ./build_kokkos/Reduced ]; then
@@ -45,6 +53,7 @@ fi
 
 
 # Build RAJA
+if $build; then
 if [ ! -d RAJA-v0.14.0 ]; then
   wget https://github.com/LLNL/RAJA/releases/download/v0.14.0/RAJA-v0.14.0.tar.gz
   tar xf RAJA-v0.14.0.tar.gz
@@ -52,6 +61,7 @@ fi
 
 cmake -H. -Bbuild_raja -DMODEL=RAJA -DRAJA_SRC=RAJA-v0.14.0 -DCMAKE_C_COMPILER=icc -DCMAKE_CXX_COMPILER=icpc -DENABLE_OPENMP=On -DRAJA_USE_COMPLEX=On
 cmake --build build_raja --parallel
+fi
 
 
 if [ -f ./build_raja/Reduced ]; then
@@ -68,8 +78,10 @@ fi
 source $HOME/intel/oneapi/setvars.sh
 source dpcpp_compiler/startup.sh
 
+if $build; then
 cmake -H. -Bbuild_sycl -DMODEL=SYCL -DCMAKE_CXX_COMPILER=clang++ -DCMAKE_CXX_FLAGS='-fsycl -fsycl-unnamed-lambda'
 cmake --build build_sycl --parallel
+fi
 
 export SYCL_DEVICE_FILTER=cpu
